refactor: criar pasta types e criar arquivo typesHome
