+++ src/components/table/TableRow.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { RowData, TableRowMethods, TableRowProps } from "../../typesTable"
+import { RowData, TableRowMethods, TableRowProps } from "../../types/typesTable"
 import themes from '../themes/table.module.scss'
 import editarImg from '../../assets/home/Edit.png'
 function TableRow(props: TableRowProps) {
