+++ src/types/typesHome.ts
@@ -1,0 +1,10 @@
+
+export  type TopicosProps = {
+    style:string
+}
+
+export type ServicosProps = {
+    imagem:string;
+    titulo:string;
+    texto:string;
+}
